feat(grpc): create service.rs and test mod
