+++ massa-grpc/src/stream/subscribe_tx_throughput.rs
@@ -1,5 +1,4 @@
-use crate::error::GrpcError;
-use crate::handler::MassaGrpcService;
+use crate::{error::GrpcError, service::MassaGrpcService};
 use futures_util::StreamExt;
 use massa_proto::massa::api::v1::{self as grpc, GetTransactionsThroughputResponse};
 use std::pin::Pin;
