+++ massa-grpc/src/tests/test.rs
@@ -1,0 +1,2 @@
+#[tokio::test]
+async fn test_grpc_server() {}
